Demo to show that Logic ignores much of the layout
Even though AU provides enough API for the host to query what kind
of layouts are possible, Logic fails to implement such a thing.

In this research project, we try to build an aumu that can be:

1. An instance with 1x stereo in, 1x stereo out.
2. An instance with 1x stereo in, 5x stereo out.
3. An instance with 1x stereo in, 1x stereo out, 8x mono out.
4. An instance with 1x stereo in, 5x stereo out, 8x mono out.

All of the following methods with their mentioned parameters have
a say in this:

* MusicDeviceBase constructor: UInt32 numInputs, UInt32 numOutputs
* bool ValidFormat: scope (or in/out), element (or bus)
* GetStreamFormat: scope (or in/out), element (or bus)
* StreamFormatWritable: scope (or in/out), element (or bus)
* GetChannelLayoutTags:  scope (or in/out), element (or bus)
* SupportedNumChannels: This is an odd one. My current take is to
  return an array with all possible global i/o channel count pairs
  that your component supports. For the above, the following would
  be returned: {{2,2},{2,10},{2,18}}. That supports all mentioned
  buses configurations. There are some tricks with negative values
  that you can do here, see Apple docs. But for enforcing a strict
  set of layouts, my current understanding is that the AU SDK
  wants us to return the value I just mentioned.
* SetBusCount: scope (or in/out), count (number of buses)
* SetAudioChannelLayout: scope, element, layout
* GetAudioChannelLayout: scope, element

These API methods and their inputs and outputs together give the
host enough resources to figure out the 4 instance configurations.

Interestingly, Logic figures out a relatively sane set of instance
configurations if we only provide what auval refers to as a
"default layout". This is done by not explicitly providing any
layouts, which in turn is done by removing GetChannelLayoutTags
and SupportedNumChannels to rely on the super class.

reuk has already analysed this before, and described in this
comment:
https://github.com/juce-framework/JUCE/issues/1508#issuecomment-2645937780

By relying on the default layout, we can at least get Logic to
offer these instances:

1. An instance with 1x stereo in, 1x stereo out.
2. An instance with 1x stereo in, 5x stereo out, 8 mono out.